add elector interface
